% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Select the most common value and return it}
\usage{
Mode(x)
}
\arguments{
\item{x}{A list of vectors.}
}
\value{
A vector.
}
\description{
Select the most common value and return it
}
\examples{
list_of_items<- c(1,2,2,3)
Mode(list_of_items)

}
